+++ docs/history.rst
@@ -10,6 +10,7 @@ Version 0.6.2 - Unreleased
 * Fixed requirements for optional ``bigtree`` dependency to ``pandas``;
 * Added field ``last_update`` to ``Device`` model;
 * Improved model admins with list display, list filters and field search;
+* Added validation on device slug from loader and fill title with slug on device creation;
 
 
 Version 0.6.1 - 2023/11/06